refactor: split + rename InventoryViewDnd → InventoryView (OLD-40)
Split the ~1000-line InventoryViewDnd.tsx into a folder-as-component:
root in index.tsx (exports InventoryView), shared types.ts, groups.ts
helpers, and per-piece files (EquippedTray, ItemContextMenu,
EncumbranceReadout, SectionCount, rows/). No behavior change.
