--- src/OscSheet/features/inventory/index.ts
@@ -1,1 +1,0 @@
-export { InventoryViewDnd } from "@features/inventory/InventoryViewDnd";
